multipass: example -> recipe
