+++ cookbook/recipes/multipass/ingredients/templates/skeleton/app/lib/multipass/types.ts
