+++ cookbook/recipes/multipass/patches/CartSummary.tsx.eb8134.patch
@@ -1,0 +1,25 @@
+index 3eae48c76..3e3f5f56e 100644
+--- a/templates/skeleton/app/components/CartSummary.tsx
++++ b/templates/skeleton/app/components/CartSummary.tsx
+@@ -12,6 +12,8 @@ import {useFetcher} from 'react-router';
+ import type {FetcherWithComponents} from 'react-router';
+ import {CartWarnings} from '~/components/CartWarnings';
+ import {CartUserErrors} from '~/components/CartUserErrors';
++// @description Import MultipassCheckoutButton for Shopify Plus multipass checkout
++import {MultipassCheckoutButton} from '~/components/MultipassCheckoutButton';
+ 
+ type CartSummaryProps = {
+   cart: OptimisticCart<CartApiQueryFragment | null>;
+@@ -58,9 +60,10 @@ function CartCheckoutActions({checkoutUrl}: {checkoutUrl?: string}) {
+ 
+   return (
+     <div>
+-      <a href={checkoutUrl} target="_self">
++      {/* @description Use MultipassCheckoutButton for Shopify Plus stores to persist customer session */}
++      <MultipassCheckoutButton checkoutUrl={checkoutUrl}>
+         <p>Continue to Checkout &rarr;</p>
+-      </a>
++      </MultipassCheckoutButton>
+       <br />
+     </div>
+   );
