+++ cookbook/recipes/multipass/patches/account.$.tsx.2df983.patch
@@ -1,0 +1,19 @@
+index 074def2a4..8ed5a1545 100644
+--- a/templates/skeleton/app/routes/account.$.tsx
++++ b/templates/skeleton/app/routes/account.$.tsx
+@@ -1,9 +1,9 @@
+ import {redirect} from 'react-router';
+ import type {Route} from './+types/account.$';
+ 
+-// fallback wild card for all unauthenticated routes in account section
+ export async function loader({context}: Route.LoaderArgs) {
+-  context.customerAccount.handleAuthStatus();
+-
+-  return redirect('/account');
+-}
++  if (await context.session.get('customerAccessToken')) {
++    return redirect('/account');
++  }
++  return redirect('/account/login');
++}
+\ No newline at end of file
