+++ cookbook/recipes/multipass/patches/account_.logout.tsx.d6592d.patch
@@ -1,0 +1,35 @@
+index 5e67cc857..d88e717d5 100644
+--- a/templates/skeleton/app/routes/account_.logout.tsx
++++ b/templates/skeleton/app/routes/account_.logout.tsx
+@@ -1,11 +1,25 @@
+-import {redirect} from 'react-router';
++import {data, redirect} from 'react-router';
+ import type {Route} from './+types/account_.logout';
+ 
+-// if we don't implement this, /account/logout will get caught by account.$.tsx to do login
++export const meta: Route.MetaFunction = () => {
++  return [{title: 'Logout'}];
++};
++
+ export async function loader() {
++  return redirect('/account/login');
++}
++
++export async function action({request, context}: Route.ActionArgs) {
++  const {session} = context;
++  session.unset('customerAccessToken');
++
++  if (request.method !== 'POST') {
++    return data({error: 'Method not allowed'}, {status: 405});
++  }
++
+   return redirect('/');
+ }
+ 
+-export async function action({context}: Route.ActionArgs) {
+-  return context.customerAccount.logout();
+-}
++export default function Logout() {
++  return null;
++}
+\ No newline at end of file
