+++ cookbook/recipes/multipass/patches/env.d.ts.899ef3.patch
@@ -1,0 +1,13 @@
+index 4b5e361ec..7dcc69f2e 100644
+--- a/templates/skeleton/env.d.ts
++++ b/templates/skeleton/env.d.ts
+@@ -5,3 +5,9 @@
+ 
+ // Enhance TypeScript's built-in typings.
+ import '@total-typescript/ts-reset';
++
++declare global {
++  interface Env {
++    PRIVATE_SHOPIFY_STORE_MULTIPASS_SECRET?: string;
++  }
++}
