+++ cookbook/recipes/multipass/patches/package.json.cfee98.patch
@@ -1,0 +1,19 @@
+index 0ee1599a1..bd532b059 100644
+--- a/templates/skeleton/package.json
++++ b/templates/skeleton/package.json
+@@ -15,6 +15,7 @@
+   "prettier": "@shopify/prettier-config",
+   "dependencies": {
+     "@shopify/hydrogen": "2025.5.0",
++    "crypto-js": "^4.2.0",
+     "graphql": "^16.10.0",
+     "graphql-tag": "^2.12.6",
+     "isbot": "^5.1.22",
+@@ -36,6 +37,7 @@
+     "@shopify/oxygen-workers-types": "^4.1.6",
+     "@shopify/prettier-config": "^1.1.2",
+     "@total-typescript/ts-reset": "^0.6.1",
++    "@types/crypto-js": "^4.2.2",
+     "@types/eslint": "^9.6.1",
+     "@types/react": "^18.2.22",
+     "@types/react-dom": "^18.2.7",
