+++ cookbook/recipes/multipass/patches/vite.config.ts.f0a2c4.patch
@@ -1,0 +1,12 @@
+index a12ad9e2e..34af7ea61 100644
+--- a/templates/skeleton/vite.config.ts
++++ b/templates/skeleton/vite.config.ts
+@@ -26,7 +26,7 @@ export default defineConfig({
+        * Include 'example-dep' in the array below.
+        * @see https://vitejs.dev/config/dep-optimization-options
+        */
+-      include: ['set-cookie-parser', 'cookie', 'react-router'],
++      include: ['set-cookie-parser', 'cookie', 'react-router', 'crypto-js'],
+     },
+   },
+ });
