--- examples/multipass/app/routes/account.$.tsx
@@ -1,8 +1,0 @@
-import {redirect, type LoaderFunctionArgs} from 'react-router';
-
-export async function loader({context}: LoaderFunctionArgs) {
-  if (await context.session.get('customerAccessToken')) {
-    return redirect('/account');
-  }
-  return redirect('/account/login');
-}
