--- examples/multipass/app/routes/account.tsx
@@ -1,205 +1,0 @@
-import {
-  data,
-  Form,
-  NavLink,
-  Outlet,
-  redirect,
-  useLoaderData,
-  type HeadersFunction,
-  type LoaderFunctionArgs,
-} from 'react-router';
-import type {CustomerFragment} from 'storefrontapi.generated';
-
-export function shouldRevalidate() {
-  return true;
-}
-
-export const headers: HeadersFunction = ({loaderHeaders}) => loaderHeaders;
-
-export async function loader({request, context}: LoaderFunctionArgs) {
-  const {session, storefront} = context;
-  const {pathname} = new URL(request.url);
-  const customerAccessToken = await session.get('customerAccessToken');
-  const isLoggedIn = !!customerAccessToken?.accessToken;
-  const isAccountHome = pathname === '/account' || pathname === '/account/';
-  const isPrivateRoute =
-    /^\/account\/(orders|orders\/.*|profile|addresses|addresses\/.*)$/.test(
-      pathname,
-    );
-
-  if (!isLoggedIn) {
-    if (isPrivateRoute || isAccountHome) {
-      session.unset('customerAccessToken');
-      return redirect('/account/login');
-    } else {
-      // public subroute such as /account/login...
-      return {
-        isLoggedIn: false,
-        isAccountHome,
-        isPrivateRoute,
-        customer: null,
-      };
-    }
-  } else {
-    // loggedIn, default redirect to the orders page
-    if (isAccountHome) {
-      return redirect('/account/orders');
-    }
-  }
-
-  try {
-    const {customer} = await storefront.query(CUSTOMER_QUERY, {
-      variables: {
-        customerAccessToken: customerAccessToken.accessToken,
-        country: storefront.i18n.country,
-        language: storefront.i18n.language,
-      },
-      cache: storefront.CacheNone(),
-    });
-
-    if (!customer) {
-      throw new Error('Customer not found');
-    }
-
-    return data(
-      {isLoggedIn, isPrivateRoute, isAccountHome, customer},
-      {
-        headers: {
-          'Cache-Control': 'no-cache, no-store, must-revalidate',
-        },
-      },
-    );
-  } catch (error) {
-    console.error('There was a problem loading account', error);
-    session.unset('customerAccessToken');
-    return redirect('/account/login');
-  }
-}
-
-export default function Acccount() {
-  const {customer, isPrivateRoute, isAccountHome} =
-    useLoaderData<typeof loader>();
-
-  if (!isPrivateRoute && !isAccountHome) {
-    return <Outlet context={{customer}} />;
-  }
-
-  return (
-    <AccountLayout customer={customer as CustomerFragment}>
-      <br />
-      <br />
-      <Outlet context={{customer}} />
-    </AccountLayout>
-  );
-}
-
-function AccountLayout({
-  customer,
-  children,
-}: {
-  customer: CustomerFragment;
-  children: React.ReactNode;
-}) {
-  const heading = customer
-    ? customer.firstName
-      ? `Welcome, ${customer.firstName}`
-      : `Welcome to your account.`
-    : 'Account Details';
-
-  return (
-    <div className="account">
-      <h1>{heading}</h1>
-      <br />
-      <AccountMenu />
-      {children}
-    </div>
-  );
-}
-
-function AccountMenu() {
-  function isActiveStyle({
-    isActive,
-    isPending,
-  }: {
-    isActive: boolean;
-    isPending: boolean;
-  }) {
-    return {
-      fontWeight: isActive ? 'bold' : undefined,
-      color: isPending ? 'grey' : 'black',
-    };
-  }
-
-  return (
-    <nav role="navigation">
-      <NavLink to="/account/orders" style={isActiveStyle}>
-        Orders &nbsp;
-      </NavLink>
-      &nbsp;|&nbsp;
-      <NavLink to="/account/profile" style={isActiveStyle}>
-        &nbsp; Profile &nbsp;
-      </NavLink>
-      &nbsp;|&nbsp;
-      <NavLink to="/account/addresses" style={isActiveStyle}>
-        &nbsp; Addresses &nbsp;
-      </NavLink>
-      &nbsp;|&nbsp;
-      <Logout />
-    </nav>
-  );
-}
-
-function Logout() {
-  return (
-    <Form className="account-logout" method="POST" action="/account/logout">
-      &nbsp;<button type="submit">Sign out</button>
-    </Form>
-  );
-}
-
-export const CUSTOMER_FRAGMENT = `#graphql
-  fragment Customer on Customer {
-    acceptsMarketing
-    addresses(first: 6) {
-      nodes {
-        ...Address
-      }
-    }
-    defaultAddress {
-      ...Address
-    }
-    email
-    firstName
-    lastName
-    numberOfOrders
-    phone
-  }
-  fragment Address on MailingAddress {
-    id
-    formatted
-    firstName
-    lastName
-    company
-    address1
-    address2
-    country
-    province
-    city
-    zip
-    phone
-  }
-` as const;
-
-// NOTE: https://shopify.dev/docs/api/storefront/latest/queries/customer
-const CUSTOMER_QUERY = `#graphql
-  query Customer(
-    $customerAccessToken: String!
-    $country: CountryCode
-    $language: LanguageCode
-  ) @inContext(country: $country, language: $language) {
-    customer(customerAccessToken: $customerAccessToken) {
-      ...Customer
-    }
-  }
-  ${CUSTOMER_FRAGMENT}
-` as const;
