--- examples/multipass/app/routes/account_.activate.$id.$activationToken.tsx
@@ -1,159 +1,0 @@
-import {
-  Form,
-  useActionData,
-  data,
-  redirect,
-  type ActionFunctionArgs,
-  type LoaderFunctionArgs,
-  type MetaFunction,
-} from 'react-router';
-
-type ActionResponse = {
-  error: string | null;
-};
-
-export const meta: MetaFunction = () => {
-  return [{title: 'Activate Account'}];
-};
-
-export async function loader({context}: LoaderFunctionArgs) {
-  if (await context.session.get('customerAccessToken')) {
-    return redirect('/account');
-  }
-  return {};
-}
-
-export async function action({request, context, params}: ActionFunctionArgs) {
-  const {session, storefront} = context;
-  const {id, activationToken} = params;
-
-  if (request.method !== 'POST') {
-    return data({error: 'Method not allowed'}, {status: 405});
-  }
-
-  try {
-    if (!id || !activationToken) {
-      throw new Error('Missing token. The link you followed might be wrong.');
-    }
-
-    const form = await request.formData();
-    const password = form.has('password') ? String(form.get('password')) : null;
-    const passwordConfirm = form.has('passwordConfirm')
-      ? String(form.get('passwordConfirm'))
-      : null;
-
-    const validPasswords =
-      password && passwordConfirm && password === passwordConfirm;
-
-    if (!validPasswords) {
-      throw new Error('Passwords do not match');
-    }
-
-    const {customerActivate} = await storefront.mutate(
-      CUSTOMER_ACTIVATE_MUTATION,
-      {
-        variables: {
-          id: `gid://shopify/Customer/${id}`,
-          input: {
-            password,
-            activationToken,
-          },
-        },
-      },
-    );
-
-    if (customerActivate?.customerUserErrors?.length) {
-      throw new Error(customerActivate.customerUserErrors[0].message);
-    }
-
-    const {customerAccessToken} = customerActivate ?? {};
-    if (!customerAccessToken) {
-      throw new Error('Could not activate account.');
-    }
-    session.set('customerAccessToken', customerAccessToken);
-
-    return redirect('/account');
-  } catch (error: unknown) {
-    if (error instanceof Error) {
-      return data({error: error.message}, {status: 400});
-    }
-    return data({error}, {status: 400});
-  }
-}
-
-export default function Activate() {
-  const action = useActionData<ActionResponse>();
-  const error = action?.error ?? null;
-
-  return (
-    <div className="account-activate">
-      <h1>Activate Account.</h1>
-      <p>Create your password to activate your account.</p>
-      <Form method="POST">
-        <fieldset>
-          <label htmlFor="password">Password</label>
-          <input
-            id="password"
-            name="password"
-            type="password"
-            autoComplete="current-password"
-            placeholder="Password"
-            aria-label="Password"
-            minLength={8}
-            required
-            // eslint-disable-next-line jsx-a11y/no-autofocus
-            autoFocus
-          />
-          <label htmlFor="passwordConfirm">Re-enter password</label>
-          <input
-            id="passwordConfirm"
-            name="passwordConfirm"
-            type="password"
-            autoComplete="current-password"
-            placeholder="Re-enter password"
-            aria-label="Re-enter password"
-            minLength={8}
-            required
-          />
-        </fieldset>
-        {error ? (
-          <p>
-            <mark>
-              <small>{error}</small>
-            </mark>
-          </p>
-        ) : (
-          <br />
-        )}
-        <button
-          className="bg-primary text-contrast rounded py-2 px-4 focus:shadow-outline block w-full"
-          type="submit"
-        >
-          Save
-        </button>
-      </Form>
-    </div>
-  );
-}
-
-// NOTE: https://shopify.dev/docs/api/storefront/latest/mutations/customeractivate
-const CUSTOMER_ACTIVATE_MUTATION = `#graphql
-  mutation customerActivate(
-    $id: ID!,
-    $input: CustomerActivateInput!,
-    $country: CountryCode,
-    $language: LanguageCode
-  ) @inContext(country: $country, language: $language) {
-    customerActivate(id: $id, input: $input) {
-      customerAccessToken {
-        accessToken
-        expiresAt
-      }
-      customerUserErrors {
-        code
-        field
-        message
-      }
-    }
-  }
-` as const;
