--- examples/multipass/app/routes/account_.register.tsx
@@ -1,211 +1,0 @@
-import {
-  Form,
-  Link,
-  useActionData,
-  data,
-  HeadersFunction,
-  redirect,
-  type ActionFunctionArgs,
-  type LoaderFunctionArgs,
-} from 'react-router';
-import type {CustomerCreateMutation} from 'storefrontapi.generated';
-
-type ActionResponse = {
-  error: string | null;
-  newCustomer:
-    | NonNullable<CustomerCreateMutation['customerCreate']>['customer']
-    | null;
-};
-
-export const headers: HeadersFunction = ({actionHeaders}) => actionHeaders;
-
-export async function loader({context}: LoaderFunctionArgs) {
-  const customerAccessToken = await context.session.get('customerAccessToken');
-  if (customerAccessToken) {
-    return redirect('/account');
-  }
-
-  return {};
-}
-
-export async function action({request, context}: ActionFunctionArgs) {
-  if (request.method !== 'POST') {
-    return data({error: 'Method not allowed'}, {status: 405});
-  }
-
-  const {storefront, session} = context;
-  const form = await request.formData();
-  const email = String(form.has('email') ? form.get('email') : '');
-  const password = form.has('password') ? String(form.get('password')) : null;
-  const passwordConfirm = form.has('passwordConfirm')
-    ? String(form.get('passwordConfirm'))
-    : null;
-
-  const validPasswords =
-    password && passwordConfirm && password === passwordConfirm;
-
-  const validInputs = Boolean(email && password);
-  try {
-    if (!validPasswords) {
-      throw new Error('Passwords do not match');
-    }
-
-    if (!validInputs) {
-      throw new Error('Please provide both an email and a password.');
-    }
-
-    const {customerCreate} = await storefront.mutate(CUSTOMER_CREATE_MUTATION, {
-      variables: {
-        input: {email, password},
-      },
-    });
-
-    if (customerCreate?.customerUserErrors?.length) {
-      throw new Error(customerCreate?.customerUserErrors[0].message);
-    }
-
-    const newCustomer = customerCreate?.customer;
-    if (!newCustomer?.id) {
-      throw new Error('Could not create customer');
-    }
-
-    // get an access token for the new customer
-    const {customerAccessTokenCreate} = await storefront.mutate(
-      REGISTER_LOGIN_MUTATION,
-      {
-        variables: {
-          input: {
-            email,
-            password,
-          },
-        },
-      },
-    );
-
-    if (!customerAccessTokenCreate?.customerAccessToken?.accessToken) {
-      throw new Error('Missing access token');
-    }
-    session.set(
-      'customerAccessToken',
-      customerAccessTokenCreate?.customerAccessToken,
-    );
-
-    return data(
-      {error: null, newCustomer},
-      {
-        status: 302,
-        headers: {
-          Location: '/account',
-        },
-      },
-    );
-  } catch (error: unknown) {
-    if (error instanceof Error) {
-      return data({error: error.message}, {status: 400});
-    }
-    return data({error}, {status: 400});
-  }
-}
-
-export default function Register() {
-  const data = useActionData<ActionResponse>();
-  const error = data?.error || null;
-  return (
-    <div className="login">
-      <h1>Register.</h1>
-      <Form method="POST">
-        <fieldset>
-          <label htmlFor="email">Email address</label>
-          <input
-            id="email"
-            name="email"
-            type="email"
-            autoComplete="email"
-            required
-            placeholder="Email address"
-            aria-label="Email address"
-            // eslint-disable-next-line jsx-a11y/no-autofocus
-            autoFocus
-          />
-          <label htmlFor="password">Password</label>
-          <input
-            id="password"
-            name="password"
-            type="password"
-            autoComplete="current-password"
-            placeholder="Password"
-            aria-label="Password"
-            minLength={8}
-            required
-          />
-          <label htmlFor="passwordConfirm">Re-enter password</label>
-          <input
-            id="passwordConfirm"
-            name="passwordConfirm"
-            type="password"
-            autoComplete="current-password"
-            placeholder="Re-enter password"
-            aria-label="Re-enter password"
-            minLength={8}
-            required
-          />
-        </fieldset>
-        {error ? (
-          <p>
-            <mark>
-              <small>{error}</small>
-            </mark>
-          </p>
-        ) : (
-          <br />
-        )}
-        <button type="submit">Register</button>
-      </Form>
-      <br />
-      <p>
-        <Link to="/account/login">Login →</Link>
-      </p>
-    </div>
-  );
-}
-
-// NOTE: https://shopify.dev/docs/api/storefront/latest/mutations/customerCreate
-const CUSTOMER_CREATE_MUTATION = `#graphql
-  mutation customerCreate(
-    $input: CustomerCreateInput!,
-    $country: CountryCode,
-    $language: LanguageCode
-  ) @inContext(country: $country, language: $language) {
-    customerCreate(input: $input) {
-      customer {
-        id
-      }
-      customerUserErrors {
-        code
-        field
-        message
-      }
-    }
-  }
-` as const;
-
-// NOTE: https://shopify.dev/docs/api/storefront/latest/mutations/customeraccesstokencreate
-const REGISTER_LOGIN_MUTATION = `#graphql
-  mutation registerLogin(
-    $input: CustomerAccessTokenCreateInput!,
-    $country: CountryCode,
-    $language: LanguageCode
-  ) @inContext(country: $country, language: $language) {
-    customerAccessTokenCreate(input: $input) {
-      customerUserErrors {
-        code
-        field
-        message
-      }
-      customerAccessToken {
-        accessToken
-        expiresAt
-      }
-    }
-  }
-` as const;
