--- examples/multipass/app/routes/cart.tsx
@@ -1,129 +1,0 @@
-import {
-  type ActionFunctionArgs,
-  data,
-  type HeadersFunction,
-  type LoaderFunctionArgs,
-  type MetaFunction,
-  useLoaderData,
-} from 'react-router';
-import type {CartQueryDataReturn} from '@shopify/hydrogen';
-import {CartForm} from '@shopify/hydrogen';
-import {CartMain} from '~/components/Cart';
-
-export const meta: MetaFunction = () => {
-  return [{title: `Hydrogen | Cart`}];
-};
-
-export const headers: HeadersFunction = ({actionHeaders}) => actionHeaders;
-
-export async function action({request, context}: ActionFunctionArgs) {
-  /***********************************************/
-  /**********  EXAMPLE UPDATE STARTS  ************/
-  const {session, cart} = context;
-
-  const [formData, customerAccessToken] = await Promise.all([
-    request.formData(),
-    session.get('customerAccessToken'),
-  ]);
-  /**********   EXAMPLE UPDATE END   ************/
-  /***********************************************/
-
-  const {action, inputs} = CartForm.getFormInput(formData);
-
-  if (!action) {
-    throw new Error('No action provided');
-  }
-
-  let status = 200;
-  let result: CartQueryDataReturn;
-
-  switch (action) {
-    case CartForm.ACTIONS.LinesAdd:
-      result = await cart.addLines(inputs.lines);
-      break;
-    case CartForm.ACTIONS.LinesUpdate:
-      result = await cart.updateLines(inputs.lines);
-      break;
-    case CartForm.ACTIONS.LinesRemove:
-      result = await cart.removeLines(inputs.lineIds);
-      break;
-    case CartForm.ACTIONS.DiscountCodesUpdate: {
-      const formDiscountCode = inputs.discountCode;
-
-      // User inputted discount code
-      const discountCodes = (
-        formDiscountCode ? [formDiscountCode] : []
-      ) as string[];
-
-      // Combine discount codes already applied on cart
-      discountCodes.push(...inputs.discountCodes);
-
-      result = await cart.updateDiscountCodes(discountCodes);
-      break;
-    }
-    case CartForm.ACTIONS.GiftCardCodesUpdate: {
-      const formGiftCardCode = inputs.giftCardCode;
-
-      // User inputted gift card code
-      const giftCardCodes = (
-        formGiftCardCode ? [formGiftCardCode] : []
-      ) as string[];
-
-      // Combine gift card codes already applied on cart
-      giftCardCodes.push(...inputs.giftCardCodes);
-
-      result = await cart.updateGiftCardCodes(giftCardCodes);
-      break;
-    }
-    case CartForm.ACTIONS.BuyerIdentityUpdate: {
-      result = await cart.updateBuyerIdentity({
-        ...inputs.buyerIdentity,
-        /***********************************************/
-        /**********  EXAMPLE UPDATE STARTS  ************/
-        customerAccessToken: customerAccessToken?.accessToken,
-        /**********   EXAMPLE UPDATE END   ************/
-        /***********************************************/
-      });
-      break;
-    }
-    default:
-      throw new Error(`${action} cart action is not defined`);
-  }
-
-  const cartId = result?.cart?.id;
-  const headers = cartId ? cart.setCartId(result.cart.id) : new Headers();
-  const {cart: cartResult, errors} = result;
-
-  const redirectTo = formData.get('redirectTo') ?? null;
-  if (typeof redirectTo === 'string') {
-    status = 303;
-    headers.set('Location', redirectTo);
-  }
-
-  return data(
-    {
-      cart: cartResult,
-      errors,
-      analytics: {
-        cartId,
-      },
-    },
-    {status, headers},
-  );
-}
-
-export async function loader({context}: LoaderFunctionArgs) {
-  const {cart} = context;
-  return await cart.get();
-}
-
-export default function Cart() {
-  const cart = useLoaderData<typeof loader>();
-
-  return (
-    <div className="cart">
-      <h1>Cart</h1>
-      <CartMain layout="page" cart={cart} />
-    </div>
-  );
-}
