--- examples/multipass/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "example-multipass",
-  "private": true,
-  "prettier": "@shopify/prettier-config",
-  "scripts": {
-    "build": "shopify hydrogen build --codegen --diff",
-    "dev": "shopify hydrogen dev --codegen --diff",
-    "preview": "shopify hydrogen preview --build --diff",
-    "lint": "eslint --no-error-on-unmatched-pattern .",
-    "typecheck": "tsc --noEmit",
-    "codegen": "shopify hydrogen codegen"
-  },
-  "dependencies": {
-    "crypto-js": "^4.2.0",
-    "snakecase-keys": "^5.5.0"
-  },
-  "devDependencies": {
-    "@types/crypto-js": "^4.2.1"
-  }
-}
